# Optimise and Refactor `Org.with_template_and_user_counts`